Move render()  & __toString() requires To RenderTrait
